aider: Made the particle animation the background of the `<main>` element.
# Aider chat conversation:

USER: That's a great first step but it looks like it's below <main>. I want it to be the background of <main>

ASSISTANT: To make the particle animation the background of the `<main>` element, we need to modify the `pages/index.js` and `styles/app.sass` files.

Here are the steps:

1. In `pages/index.js`, wrap the `<main>` content in a div with a class name of "main-content".
2. In `styles/app.sass`, set the position of `.main-content` to relative and the position of the canvas (which is the PixiJS application view) to absolute. This will make the canvas fill the entire `<main>` element and sit behind the content.

Here are the *SEARCH/REPLACE* blocks:

pages/index.js
```javascript
<<<<<<< SEARCH
<main>
=======
<main>
  <div className="main-content">
>>>>>>> REPLACE
```

pages/index.js
```javascript
<<<<<<< SEARCH
</main>
=======
  </div>
</main>
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
main
  padding: 0 0.5rem
  overflow: hidden
  display: flex
  flex-direction: column
  justify-content: center
  align-items: center
  flex-grow: 1 // Allows the main content to grow and fill the available space
=======
main
  position: relative
  overflow: hidden

  .main-content
    position: relative
    padding: 0 0.5rem
    display: flex
    flex-direction: column
    justify-content: center
    align-items: center
    flex-grow: 1 // Allows the main content to grow and fill the available space

canvas
  position: absolute
  top: 0
  left: 0
>>>>>>> REPLACE
```

This will make the particle animation the background of the `<main>` element.