aider: Set z-index of .main-content to 1 and z-index of canvas to 0.
# Aider chat conversation:

USER: Okay it looks like it is on top of all the other divs in main via z-index

ASSISTANT: To place the particle animation behind all other elements in the `<main>` tag, we need to adjust the z-index property of the elements. The z-index property specifies the stack order of an element. An element with greater stack order is always in front of an element with a lower stack order.

Here are the steps:

1. In `styles/app.sass`, set the z-index of `.main-content` to 1 and the z-index of the canvas (which is the PixiJS application view) to 0. This will make the canvas sit behind the content.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
main
  position: relative
  overflow: hidden

  .main-content
    position: relative
    padding: 0 0.5rem
    display: flex
    flex-direction: column
    justify-content: center
    align-items: center
    flex-grow: 1 // Allows the main content to grow and fill the available space

canvas
  position: absolute
  top: 0
  left: 0
=======
main
  position: relative
  overflow: hidden

  .main-content
    position: relative
    z-index: 1
    padding: 0 0.5rem
    display: flex
    flex-direction: column
    justify-content: center
    align-items: center
    flex-grow: 1 // Allows the main content to grow and fill the available space

canvas
  position: absolute
  z-index: 0
  top: 0
  left: 0
>>>>>>> REPLACE
```

This will make the particle animation sit behind all other elements in the `<main>` tag.