aider: Adjusted footer and main content CSS properties to ensure the footer is always visible at the bottom of the page and the main content is scrollable if necessary.
# Aider chat conversation:

USER: Awesome, I think we need to do something with the footer too because I know longer see it

ASSISTANT: It seems like the footer might be hidden behind other elements due to the z-index property or it might be pushed out of the viewport due to the height of other elements.

To fix this, we can adjust the CSS properties of the footer and the main content. We can give the main content a specific height and make it scrollable if the content overflows. This will ensure that the footer is always visible at the bottom of the page.

Here are the *SEARCH/REPLACE* blocks:

styles/app.sass
```sass
<<<<<<< SEARCH
main
  position: relative
  overflow: hidden

  .main-content
    position: relative
    z-index: 1
    padding: 0 0.5rem
    display: flex
    flex-direction: column
    justify-content: center
    align-items: center
    flex-grow: 1 // Allows the main content to grow and fill the available space
=======
main
  position: relative
  overflow: auto // Make main scrollable if content overflows

  .main-content
    position: relative
    z-index: 1
    padding: 0 0.5rem
    display: flex
    flex-direction: column
    justify-content: center
    align-items: center
    height: calc(100% - 60px) // Deduct the height of the footer
    overflow: auto // Make main-content scrollable if content overflows
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
footer
  position: absolute
  bottom: 0
  margin-top: 0.25rem
  margin-bottom: 0.25rem
  width: 100%
  display: flex
  justify-content: center
  align-items: center
=======
footer
  position: relative // Change from absolute to relative
  height: 60px // Specify a height for the footer
  width: 100%
  display: flex
  justify-content: center
  align-items: center
>>>>>>> REPLACE
```

Please apply these changes and let me know if the footer is now visible.