aider: Added styles to hide Quora and StackOverflow icons on mobile devices.
# Aider chat conversation:

ASSISTANT: Great, now that we have the correct file path for the SASS file, we can proceed to add the necessary styles to hide the Quora and StackOverflow icons on mobile devices.

Here are the *SEARCH/REPLACE* blocks for the `styles/app.sass` file:

/home/smusso/code/Blog/styles/app.sass
```sass
<<<<<<< SEARCH
@media (max-width: 800px) // mobile specific styles
=======
@media (max-width: 800px) // mobile specific styles
  #quora-link, #stackoverflow-link
    display: none
>>>>>>> REPLACE
```

This change will hide the Quora and StackOverflow icons when the website is viewed on devices with a screen width of 800 pixels or less.