aider: Add CSS class to profile image in header to change cursor to pointer when hovered over.
# Aider chat conversation:

USER: When I hover over the profile image in the header my cursor should indicate it's a link

ASSISTANT: To make this change, we need to modify `components/Header.js`:

1. Add a CSS class to the image that changes the cursor to a pointer when hovered over.

Here is the *SEARCH/REPLACE* block:

components/Header.js
```javascript
<<<<<<< SEARCH
<Image
  src="/images/headshot.jfif"
  alt="Headshot"
  width={50}
  height={50}
  onClick={() => router.push('/')}
/>
=======
<Image
  src="/images/headshot.jfif"
  alt="Headshot"
  width={50}
  height={50}
  onClick={() => router.push('/')}
  className="cursor-pointer"
/>
>>>>>>> REPLACE
```

Please note that the CSS class `cursor-pointer` should be defined in your CSS file with the property `cursor: pointer;`. If it's not already defined, you'll need to add it.