aider: Add support for using the `Image` component in MDX files.
# Aider chat conversation:

ASSISTANT: The `Image` component from `next/image` is a React component and can't be used directly in MDX or Markdown files. However, you can use it in your MDX files by providing it as a component in the scope when rendering the MDX content.

In your `pages/blog/[slug].js` file, you can provide the `Image` component in the components prop of the `MDXRemote` component. Here's how you can do it:

```javascript
import Image from 'next/image'
import { MDXRemote } from 'next-mdx-remote'

// ...

<MDXRemote {...mdxSource} components={{ Image }} />
```

Then, in your MDX content, you can use the `Image` component like this:

```markdown
<Image src="/path/to/image.jpg" alt="description of image" width={500} height={300} />
```

Remember to replace `"/path/to/image.jpg"` with the path to your image, and `"description of image"` with a description of the image. The `width` and `height` props are required and they set the layout dimensions of the image.

Here are the *SEARCH/REPLACE* blocks:

pages/blog/[slug].js
```javascript
<<<<<<< SEARCH
<MDXRemote {...mdxSource} />
=======
<MDXRemote {...mdxSource} components={{ Image }} />
>>>>>>> REPLACE
```

This will allow you to use the `Image` component in your MDX files. However, you'll need to ensure that the image paths are correct and the images are available in the public directory of your Next.js project.