aider: Wrap MDXRemote component in a div with the id "mdx-body".
# Aider chat conversation:

ASSISTANT: The `mdxOptions` object passed to the `serialize` function is used to configure the MDX compiler. It can include `remarkPlugins` and `rehypePlugins` which are used to transform the MDX content during compilation. However, it doesn't provide a direct way to add a CSS id to the 'body' of the MDX content.

The 'body' of the MDX content is essentially the content of the MDX file that is being rendered by the `MDXRemote` component. To add a CSS id to this content, you would typically wrap the `MDXRemote` component in a parent element with the desired id.

Here's an example of how you can do this:

pages/blog/[slug].js
```javascript
<<<<<<< SEARCH
return <MDXRemote {...mdxSource} components={{ Image }}/>
=======
return (
  <div id="mdx-body">
    <MDXRemote {...mdxSource} components={{ Image }}/>
  </div>
);
>>>>>>> REPLACE
```

In this example, the `MDXRemote` component is wrapped in a `div` with the id "mdx-body". You can replace "mdx-body" with the id you want to use.

Please note that this will add the id to the parent element of the MDX content, not to the MDX content itself. If you need to add the id to a specific part of the MDX content, you would need to do this in the MDX file itself.