aider: Add text and image wrapping on the `About` page.
# Aider chat conversation:

ASSISTANT: To implement text wrapping / image wrapping on the `About` page, we need to modify `pages/about/index.js` and `styles/app.sass`:

1. In `pages/about/index.js`, we need to add a new class to the image element and adjust the structure of the paragraphs to allow for text wrapping.
2. In `styles/app.sass`, we need to add new styles for the image class to float the image to the right and allow text to wrap around it.

Here are the *SEARCH/REPLACE* blocks:

pages/about/index.js
```javascript
<<<<<<< SEARCH
        <p className="">
          I'm your host, Sam Musso. I'm a software engineer, a polyglot, and a
          tinkerer.
          <img src="/static/sam-musso.jpg" alt="Sam Musso" id="img-bio" />
        </p>
=======
        <div className="bio-section">
          <img src="/static/sam-musso.jpg" alt="Sam Musso" id="img-bio" className="bio-image" />
          <p className="">
            I'm your host, Sam Musso. I'm a software engineer, a polyglot, and a
            tinkerer.
          </p>
        </div>
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
=======
.bio-section
  display: flex
  flex-direction: row
  justify-content: start
  align-items: start

.bio-image
  float: right
  margin: 0 0 1em 1em
>>>>>>> REPLACE
```

This will float the image to the right of the text and allow the text to wrap around it. The `bio-section` class is used to contain the image and the paragraph, and the `bio-image` class is used to style the image. The `margin` property in the `bio-image` class creates space around the image.