aider: Add styles to control the visibility and transition of the images in the `.bio-image` class.
# Aider chat conversation:

ASSISTANT: To complete the feature, we need to modify the `styles/app.sass` file to add styles that will control the visibility and transition of the images.

We will add styles to make the `.bio-image` class have an opacity of 0 by default, and then add styles to make the `.active` class have an opacity of 1. We will also add a transition property to the `.bio-image` class to control the transition effect when the image becomes active.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
.bio-image
  float: right
  margin: 0 0 1em 1em
  clear: both
=======
.bio-image
  float: right
  margin: 0 0 1em 1em
  clear: both
  opacity: 0
  transition: opacity 1s

.bio-image.active
  opacity: 1
>>>>>>> REPLACE
```