aider: To make the images overlap, the styles in `styles/app.sass` have been modified to position the images absolutely within the `.image-container` div.
# Aider chat conversation:

USER: That's a great first cut! Okay the first refinment we need to make is that right now the images are stacked on top of each other. We need it so that the images overlap on the dom

ASSISTANT: To make the images overlap, we need to modify the `styles/app.sass` file to add styles that will position the images absolutely within the `.image-container` div.

We will add styles to make the `.image-container` class have a relative position, and then add styles to make the `.bio-image` class have an absolute position. This will make the images overlap each other.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
.bio-image
  float: right
  margin: 0 0 1em 1em
  clear: both
  opacity: 0
  transition: opacity 1s
=======
.image-container
  position: relative

.bio-image
  position: absolute
  top: 0
  right: 0
  opacity: 0
  transition: opacity 1s
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
.bio-image.active
  opacity: 1
=======
.bio-image.active
  opacity: 1
  z-index: 1
>>>>>>> REPLACE
```