aider: Refactored image loading logic to use Promise.all for ensuring both images are loaded before drawing.
# Aider chat conversation:

USER: Hmm I don't see images on the canvas anymore, it's blank

ASSISTANT: The issue might be related to the loading of the images. The `onload` event is set only for the first image, which means the second image might not be loaded when the `draw` function is called.

To fix this, we need to ensure that both images are loaded before starting the drawing process. We can do this by using `Promise.all` to wait for both images to load.

Here are the steps to modify `pages/about/index.js`:

1. Replace the current image loading logic with a `Promise.all` that waits for both images to load.
2. Move the setting of the canvas dimensions and context into the `Promise.all` callback.

Here are the *SEARCH/REPLACE* blocks:

pages/about/index.js
```javascript
<<<<<<< SEARCH
    const images = [new Image(), new Image()];

    images[0].src = "/static/sam-musso.jpg";
    images[1].src = "/static/sam-solar.png";

    images[0].onload = function () {
      const canvas = canvasRef.current;
      canvas.width = this.naturalWidth;
      canvas.height = this.naturalHeight;
      setContext(canvas.getContext("2d"));
    };
=======
    const loadImages = [
      new Promise((resolve) => {
        const img = new Image();
        img.onload = () => resolve(img);
        img.src = "/static/sam-musso.jpg";
      }),
      new Promise((resolve) => {
        const img = new Image();
        img.onload = () => resolve(img);
        img.src = "/static/sam-solar.png";
      }),
    ];

    Promise.all(loadImages).then((images) => {
      const canvas = canvasRef.current;
      canvas.width = images[0].naturalWidth;
      canvas.height = images[0].naturalHeight;
      setContext(canvas.getContext("2d"));
    });
>>>>>>> REPLACE
```

This change will ensure that both images are loaded before the drawing process starts.