aider: Moved the `images` variable to a higher scope to fix the ReferenceError.
# Aider chat conversation:

USER: Unhandled Runtime Error
ReferenceError: images is not defined

Source
pages/about/index.js (62:26) @ images

ASSISTANT: The error is due to the `images` variable being out of scope. It is defined inside the `Promise.all(loadImages).then((images) => {...})` block, but it is being accessed outside of this block in the `draw` function.

To fix this, we need to move the `images` variable to a higher scope so it can be accessed by both the `Promise.all` block and the `draw` function.

Here is the *SEARCH/REPLACE* block:

pages/about/index.js
```javascript
<<<<<<< SEARCH
  useEffect(() => {
    const loadImages = [
      new Promise((resolve) => {
        const img = new Image();
        img.onload = () => resolve(img);
        img.src = "/static/sam-musso.jpg";
      }),
      new Promise((resolve) => {
        const img = new Image();
        img.onload = () => resolve(img);
        img.src = "/static/sam-solar.png";
      }),
    ];

    Promise.all(loadImages).then((images) => {
      const canvas = canvasRef.current;
      canvas.width = images[0].naturalWidth;
      canvas.height = images[0].naturalHeight;
      setContext(canvas.getContext("2d"));
    });
=======
  const [images, setImages] = useState([]);

  useEffect(() => {
    const loadImages = [
      new Promise((resolve) => {
        const img = new Image();
        img.onload = () => resolve(img);
        img.src = "/static/sam-musso.jpg";
      }),
      new Promise((resolve) => {
        const img = new Image();
        img.onload = () => resolve(img);
        img.src = "/static/sam-solar.png";
      }),
    ];

    Promise.all(loadImages).then((loadedImages) => {
      const canvas = canvasRef.current;
      canvas.width = loadedImages[0].naturalWidth;
      canvas.height = loadedImages[0].naturalHeight;
      setContext(canvas.getContext("2d"));
      setImages(loadedImages);
    });
>>>>>>> REPLACE
```

This change will ensure that the `images` variable is accessible in the `draw` function.