aider: Add style rule to make footer text display as block-level elements.
# Aider chat conversation:

ASSISTANT: To make the footer text appear on separate lines, we need to modify the CSS rules for the `p` tag within the `footer` in `styles/app.sass`.

Currently, the `p` tag within the `footer` does not have any specific styles applied to it. We need to add a style rule to make the text within the `p` tag display as block-level elements, which will cause each `p` tag to start on a new line.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
footer
  position: absolute
  bottom: 0
  width: 100%
  height: 60px
  display: flex
  justify-content: center
  align-items: center
  z-index: 100
=======
footer
  position: absolute
  bottom: 0
  width: 100%
  height: 60px
  display: flex
  justify-content: center
  align-items: center
  z-index: 100

  p
    display: block
>>>>>>> REPLACE
```