aider: Set footer text to break into a new line by changing flex-direction to column.
# Aider chat conversation:

USER: Hmm no it's still on one line

ASSISTANT: I see. It seems like the footer text is not breaking into a new line as expected. This could be due to the flex property in the footer style. Flex items are, by default, displayed in a row (horizontally). We can change this behavior by setting the flex-direction property to column.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
footer
  position: absolute
  bottom: 0
  width: 100%
  height: 60px
  display: flex
  justify-content: center
  align-items: center
  z-index: 100

  p
    display: block
=======
footer
  position: absolute
  bottom: 0
  width: 100%
  height: 60px
  display: flex
  flex-direction: column
  justify-content: center
  align-items: center
  z-index: 100

  p
    display: block
>>>>>>> REPLACE
```